Feat: Add __ADDONLY__ flag on GroupMap
By adding an __ADDONLY__ flag in GroupMap, it is possible to allow SAML
attributes to add a wiki-group to a user but make the lack of such SAML
attributes not remove the group. This allows automatic adding to groups
for users who have accurate SAML attributes, but enables manual setting
of wiki-groups for users with inaccurate or incomplete SAML attributes.

Change-Id: Ieef04bfb554b47c2d8e6f0aeb44fd2dcd4414fe6